Tracy Profiler (#88)
* Direct move fix

* Fixed handling of hashes in dict/array properties

* Dialogs sex tag fix

* Skip processing of destroyed critters

* Embedded resources reading fix

* World map fixes

* prepare-workspace excluded from validate script

* Linux binaries packaging

* Embedded data set to vloatile

* SIGINT/SIGTERM handling in headless mode

* Manual process forking by passing --fork

* Dummy window for NullRenderer

* ForkProcess to GenericUtils

* ForkProcess to GenericUtils

* Correct default embedded font

* Not initialized client exit fix

* Server loading split to initialization and logic start

* Improved packs network updating speed

* Command line settings now applied after all

* DataSynchronization fix in test runs

* Sprite manager fix

* GetExePath for future use

* Compilation fix

* Compilation fix

* fork tuning

* fork tuning

* Bakering and packaging fixes and tuning

* Updater loading screen

* Compilation fix

* Compilation fix

* SPIRV-Cross updated to 1.3.236.0

* glslang updated to 11.13.0/1.3.236.0

* FBX SDK updated to 2020.3.1

* Revert "FBX SDK updated to 2020.3.1"

This reverts commit 5a7ba295409f62ddb65478ef163073cb3f5c5e85.

* Clang-cl compilation fix

* Logging tuning

* Build tuning

* Restored saved deferred calls

* Server initialization fixes and tuning

* Reworked data loading (location -> map -> critter/item -> item)

* Mouse input tuning + revisited and applied necessary SDL hints

* Compilation fix

* Properties historical data now writed to main data base

* Client-server communication fixes

* External configs loading fix

* Server-*.bat Client-*.bat Mapper-*.bat without bakering

* Enabled mouse auto capturing

* Excluded verbose stack trace, now only brief

* Static analyse fix

* Removed OnCritterGlobalMapIdle event

* Compatability version temporary will be unique per commit

* SDL updated to 2.26.2

* Restored Web packaging

* SDL readme

* Outdated resources checks

* Log time zero padding

* Files renaming

* Typo fix

* FO_DEV_NAME to android log section

* Data sync fixes

* Todo done

* Critter 3d model fixes

* Compilation fix

* Mouse targeting tuning

* GCC compilation fix

* Mongo driver disabled GCC automatic init/cleanup

* Tracy Profiler 0.9

* Readme

* CMake connection

* Tracy include and frame mark

* All functions marked by PROFILER_ENTRY

* Tracy enabled in RelWithDebInfo build + renaming

* Renaming

* Manual stack trace collection

* Compilation fixes

* Formatting fix

* Debugger friendly call stack storing

* LTO for all

* Start server with delay

* Added AngelScript script function calls to Tracy

* Cleanup code and todo'es

+++ Source/Common/StringUtils.cpp
@@ -38,9 +38,9 @@
 
 // ReSharper disable CppInconsistentNaming
 
-auto _str::length() const -> uint
+auto _str::length() const -> size_t
 {
-    return static_cast<uint>(_s.length());
+    return _s.length();
 }
 
 auto _str::empty() const -> bool
@@ -104,9 +104,9 @@ auto _str::isValidUtf8() const -> bool
     return true;
 }
 
-auto _str::lengthUtf8() const -> uint
+auto _str::lengthUtf8() const -> size_t
 {
-    uint length = 0;
+    size_t length = 0;
     const auto* str = _s.c_str();
 
     while (*str != 0) {
